+++ dpws/src/main/java/org/somda/sdc/dpws/client/DiscoveryFilter.java
@@ -1,5 +1,8 @@
 package org.somda.sdc.dpws.client;
 
+import org.somda.sdc.dpws.soap.wsdiscovery.MatchBy;
+
+import javax.annotation.Nullable;
 import javax.xml.namespace.QName;
 import java.util.Collection;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -13,6 +16,10 @@ public class DiscoveryFilter {
 
     private final Collection<QName> types;
     private final Collection<String> scopes;
+
+    @Nullable
+    private final MatchBy matchBy;
+
     private final String discoveryId;
 
 
@@ -25,8 +32,23 @@ public class DiscoveryFilter {
      * >WS-Discovery Probe</a>
      */
     public DiscoveryFilter(Collection<QName> types, Collection<String> scopes) {
+        this(types, scopes, null);
+    }
+
+    /**
+     * Creates a new discovery filter with a discovery id that is unique across one application instance.
+     *
+     * @param types  the types to match.
+     * @param scopes the scopes to match.
+     * @param matchBy the scopes matching rule, or null to use the default
+     *                <code>http://docs.oasis-open.org/ws-dd/ns/discovery/2009/01/rfc3986</code>.
+     * @see <a href="http://docs.oasis-open.org/ws-dd/discovery/1.1/os/wsdd-discovery-1.1-spec-os.html#_Toc234231831"
+     * >WS-Discovery Probe</a>
+     */
+    public DiscoveryFilter(Collection<QName> types, Collection<String> scopes, @Nullable MatchBy matchBy) {
         this.types = types;
         this.scopes = scopes;
+        this.matchBy = matchBy;
         this.discoveryId = Integer.toString(DISCOVERY_ID_COUNTER.incrementAndGet());
     }
 
@@ -36,6 +58,10 @@ public class DiscoveryFilter {
 
     public Collection<String> getScopes() {
         return scopes;
+    }
+
+    public MatchBy getMatchBy() {
+        return matchBy;
     }
 
     /**